feat(APIStats): add reorder and deleting filters - Also groups APIStats components into directory to avoid APIStats file too large
