+++ package.json
@@ -9,6 +9,7 @@
     "@emotion/css": "^11.10.5",
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
+    "@mui/icons-material": "^5.11.0",
     "@mui/material": "^5.10.16",
     "@mui/x-data-grid": "^5.17.14",
     "@testing-library/jest-dom": "^5.16.5",
