+++ src/pages/APIStats/index.tsx
@@ -1,0 +1,2 @@
+import APIStats from './APIStats';
+export default APIStats;
